+++ examples/advanced/src/gen/models/ts/Customer.ts
@@ -1,5 +1,13 @@
 import type { Address } from './Address.ts'
 
+export const customerParamsStatusEnum = {
+  placed: 'placed',
+  approved: 'approved',
+  delivered: 'delivered',
+} as const
+
+export type CustomerParamsStatusEnumKey = (typeof customerParamsStatusEnum)[keyof typeof customerParamsStatusEnum]
+
 export type Customer = {
   /**
    * @type integer | undefined, int64
@@ -9,6 +17,20 @@ export type Customer = {
    * @type string | undefined
    */
   username?: string
+  /**
+   * @type object | undefined
+   */
+  params?: {
+    /**
+     * @description Order Status
+     * @type string
+     */
+    status: CustomerParamsStatusEnumKey
+    /**
+     * @type string
+     */
+    type: string
+  }
   /**
    * @type array | undefined
    */