+++ examples/advanced/src/gen/zod/customerSchema.ts
@@ -6,6 +6,12 @@ import { addressSchema } from './addressSchema.ts'
 export const customerSchema = z.object({
   id: z.optional(z.number().int()),
   username: z.optional(z.string()),
+  params: z.optional(
+    z.object({
+      status: z.enum(['placed', 'approved', 'delivered']).describe('Order Status'),
+      type: z.string(),
+    }),
+  ),
   address: z.optional(z.array(z.lazy(() => addressSchema))),
 }) as unknown as ToZod<Customer>