+++ examples/advanced/src/gen/zod/orderSchema.ts
@@ -5,6 +5,12 @@ import type { Order } from '../models/ts/Order.ts'
 export const orderSchema = z.object({
   id: z.optional(z.number().int().min(3).max(100)),
   petId: z.optional(z.number().int()),
+  params: z.optional(
+    z.object({
+      status: z.enum(['placed', 'approved', 'delivered']).describe('Order Status'),
+      type: z.string(),
+    }),
+  ),
   quantity: z.optional(z.number().int()),
   orderType: z.optional(z.enum(['foo', 'bar'])),
   type: z.optional(z.string().describe('Order Status')),