+++ packages/plugin-faker/src/plugin.ts
@@ -27,6 +27,7 @@ export const pluginFaker = definePlugin<PluginFaker>((options) => {
     contentType,
   } = options
 
+  // @deprecated Will be removed in v5 when collisionDetection defaults to true
   const usedEnumNames = {}
 
   return {