+++ packages/plugin-react-query/src/components/Mutation.tsx
@@ -2,7 +2,7 @@ import { isOptional, type Operation } from '@kubb/oas'
 import type { OperationSchemas } from '@kubb/plugin-oas'
 import { getComments, getPathParams } from '@kubb/plugin-oas/utils'
 import { File, Function, FunctionParams } from '@kubb/react-fabric'
-import type { KubbNode } from '@kubb/react-fabric/types'
+import type { FabricReactNode } from '@kubb/react-fabric/types'
 import type { PluginReactQuery } from '../types.ts'
 import { MutationKey } from './MutationKey.tsx'
 import { MutationOptions } from './MutationOptions.tsx'
@@ -82,7 +82,7 @@ export function Mutation({
   operation,
   mutationKeyName,
   customOptions,
-}: Props): KubbNode {
+}: Props): FabricReactNode {
   const mutationKeyParams = MutationKey.getParams({
     pathParamsType,
     typeSchemas,