+++ packages/plugin-ts/src/generators/typeGenerator.test.tsx
@@ -557,7 +557,7 @@ describe('typeGenerator schema', async () => {
       output: './gen',
     })
 
-    const schemas = getSchemas({ oas })
+    const { schemas } = getSchemas({ oas })
     const name = props.path
     const schema = schemas[name] as SchemaObject
     const tree = generator.parse({ schema, name, parentName: null })