+++ tests/3.0.x/main.test.ts
@@ -136,6 +136,58 @@ const configs: Array<{ name: string; config: UserConfig }> = [
       ],
     },
   },
+  {
+    name: 'caseSensitivity',
+    config: {
+      root: __dirname,
+      input: {
+        path: '../../schemas/3.0.x/caseSensitivity.yaml',
+      },
+      output: {
+        path: './gen',
+        barrelType: false,
+      },
+      plugins: [
+        pluginOas({
+          validate: false,
+          generators: [],
+          collisionDetection: true,
+        }),
+        pluginTs({
+          output: {
+            path: './types',
+            barrelType: false,
+          },
+        }),
+      ],
+    },
+  },
+  {
+    name: 'duplicateEnum',
+    config: {
+      root: __dirname,
+      input: {
+        path: '../../schemas/3.0.x/duplicateEnum.yaml',
+      },
+      output: {
+        path: './gen',
+        barrelType: false,
+      },
+      plugins: [
+        pluginOas({
+          validate: false,
+          generators: [],
+          collisionDetection: true,
+        }),
+        pluginTs({
+          output: {
+            path: './types',
+            barrelType: false,
+          },
+        }),
+      ],
+    },
+  },
 ]
 
 describe(`Main OpenAPI ${version}`, () => {