+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @kubb/cli
 
+## 4.19.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @kubb/core@4.19.2
+
+## 4.19.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @kubb/core@4.19.1
+
+## 4.19.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @kubb/core@4.19.0
+
 ## 4.18.5
 
 ### Patch Changes